Introduce new plugin for menus with roving tab index
Add plugin directory with menu plugin that uses a PluginView render a set of menu buttons above the editor to issue commands for formatting content.
The PluginView ensures that the menu buttons update with the current cursor position.
Buttons are tab-able and disable in different modes for now. We will look at spiking toggle-ability later.
This still causes a minor issue where if the user tabs back to the menu and the button has been disabled then it jumps back to the first non-disabled button which might be a bit weird. Will double check what behaviour we want.

Co-authored-by: Daniel Karaj <daniel.karaj@digital.cabinet-office.gov.uk>
Co-authored-by: Ryan Brown <ryan.brown@digital.cabinet-office.gov.uk>