Added playwright tests for E2E testing
Rather than indivisually testing components of the page, might make more sense to test the wider functionality of each custom node and menu item.
The way we test nodes might differ in case if it is inline editing or editing blocks of text. Splitting them into these categories.
For each existing menu item, test happy paths that the element is:
- loadable
- renders expected menu
- able to type in mode
- able to turn existing text into said mode
- able to toggle it off or embed other elements within the mode

Co-authored-by: Daniel Karaj <daniel.karaj@digital.cabinet-office.gov.uk>
Co-authored-by: Ryan Brown <ryan.brown@digital.cabinet-office.gov.uk>