Stop importing the prosemirror markdown nodes
Copy node schema definitions from the prosemirror markdown codebase to the node definitions defined in visual editor, excluding code blocks, horizontal rules and h1s from the schema.

This makes the way in which all nodes are defined consistent, which will make future changes easier and remove markdown features that we will not be supporting.

Lists are copied from the `prosemirror-schema-list` instead, as before.