Stop importing the prosemirror markdown marks
Copy the link mark schema definition from the prosemirror markdown codebase to a new marks directory.

This brings marks in line with nodes, will make future changes easier and removes markdown features that we will not be supporting.

Update the menu plugin view to handle having no focusable buttons and remove references to bold in the tests.