Make tests for editor toolbar state reflect intended behaviour
These tests are intended to ensure that we have the correct toolbar elements enabled or disabled for each node context.

Because they were only testing that toolbar elements were visible, we were not asserting that some toolbar elements were disabled. Asserting that certain toolbar elements are disabled is a valuable property of the test suite, because that is the only way we can confirm that the Prosemirror schema is configured as intended.

This commit solves this problem by asserting that each toolbar element is either enabled or disabled, as appropriate