Move toolbar definition out of nodes into menu.js
With upcoming changes to the menu, the mapping of nodes to menu items doesn't hold and the order needs to be explicitly defined. This commit removes the menu item definition from nodes and defines new buttons and command in menu.js to centralise menu item definition.
It uses unicode icons for the buttons - to be changed to SVGs in a follow up PR and ProseMirror's `chainCommands` to allow the h2 button to toggle.