Update tests to reflect new menu
FIxes tests that feature buttons that are present in the new menu.
Deletes tests for paragraph button the behaviour is now covered in the heading button.
Amends tests that cover components that will appear in the dropdowns with `fixme` so that they can be fixed when those components are reintroduced.