Add custom keymapping for addresses
When in an address, insert a hard break instead of a paragraph.
When at the end of an address, replace the hard break with a paragraph and then lift it out of the address block.

Add testing for the enter key behaviour by checking the  markdown output.

There is a bit of a weird behaviour for Address where GovSpeak removes the last trailing space for GovSpeak output, but still visual editor still renders it with a space - but this shouldn't be a major setback for the user.