Reuse workflow for publishing package
Extracted the publishing of the package into a reusable workflow in
govuk-infrastructure repo. Pulling that in and passing through the
ALPHAGOV_NPM_AUTOMATION_TOKEN as a param. The name of the package
default to the name of the repo. Setting the deploy_to_github_pages
to false as we have not set that up yet.