Add markdown output testing for the e2e tests
Adding a e2e check where it should be rendering expected GovSpeak.
There is a bit of a weird behaviour for Address where GovSpeak removes the last trailing space for GovSpeak output, but still visual editor still renders it with a space - but this shouldn't be a major setback for the user.