Make tabbing behaviour consistent across browsers
We're overriding the default keyup behaviour which, in Firefox,
is causing the arrow right and left press to select a drop down
rather than move right or left to the neighboring menu item.

With the current changes, arrow right and left are consistently
tabbing through the menu bar in all browsers we care about.

This roving tab index behaviour is in line with accessibility
guidelines. We chose to follow the guidelines rather than
preserve the default Firefox behaviour which traverses down
the drop-down items when using Left and Right arrow keys.
See https://www.w3.org/WAI/ARIA/apg/patterns/toolbar/examples/toolbar.