Improve deletion of list items
Simplify the custom list item deletion to simply joinBackwards twice - once escaping the list item and a second time to join the paragraph with the previous list item.

This fixes two cases (which now have tests) where it was impossible to delete an empty list and that the custom list deletion didn't apply when the cursor was at the beginning of a list item that wasn't empty.