Simulate paste events for firefox.
Triggering the browser clipboard functionality using keyboard shortcuts doesn't work in headless chrome (i.e. when tests are run in github actions).
This means that the images tests for paste parsing were always failing.
By simulating the events instead of we avoid that issue.

This approach unfortunately does not work in firefox, so we keep the old approach there